pyproject.toml
[tool.poetry]
name = "stocker"
version = "0.1.0"
description = ""
authors = ["Nikolaus Gullotta <nikolaus.gullotta@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.2"
yfinance = "^0.2.3"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
autoflake = "^2.0.0"
pytest = "^7.2.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 79